test(checkout): CHECKOUT-9079 convert coupon and promotion tests from enzyme to rtl
